Rename `block` package to `header`
