+++ relay/pkg/header/forwarder.go
@@ -1,4 +1,4 @@
-package block
+package header
 
 import (
 	"context"
